backend/requirements.txt
asgiref==3.8.1
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.2
djangorestframework==3.16.0
et_xmlfile==2.0.0
lxml==5.4.0
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pycparser==2.22
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
